+++ src/checkout/types/getCheckout.ts
@@ -154,6 +154,7 @@ export interface getCheckout_checkout_lines_variant_product {
 
 export interface getCheckout_checkout_lines_variant {
   __typename: "ProductVariant";
+  stockQuantity: number;
   id: string;
   name: string;
   price: getCheckout_checkout_lines_variant_price | null;